requirements.txt
asgiref==3.2.3
cmake==3.16.3
cycler==0.10.0
decorator==4.4.1
Django==3.0.3
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
dlib==19.18.0
imageio==2.6.1
kiwisolver==1.1.0
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
pyparsing==2.4.6
pypiwin32==223
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
